+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.1 (unreleased)
+
+- Fix issue with idempotency when multiple devices are used
+
 ## 0.4.0
 
 - Add option to manage child objects with `nxos_rest` resource